Split out adding YouTube and self-hosted assets
